Installation of initial QPFHMI.ini file if not existing
git-svn-id: https://euclid.esac.esa.int/svn/ESA/SOC/SOC-3-DEV/SOC-3-07-QLook/QPF/branches/dev.HMIalone@14553 463500fe-d6be-4184-9b20-47717e75ecac

+++ scripts/BuildQPF.sh
@@ -268,8 +268,10 @@ done
 install_scpt RunQPF.sh
 
 QPF_INI="${RUN_PATH}/QPFHMI.conf"
-mkdir -p ${HOME}/.config/QPF
-cp "${QPF_INI}" ${HOME}/.config/QPF
+if [ ! -f "${HOME}/.config/QPF/${QPF_INI}" ]; then
+    mkdir -p ${HOME}/.config/QPF
+    cp "${QPF_INI}" ${HOME}/.config/QPF
+fi
 
 ## Creating QPFDB database
 step "Setting up QPF database"
